Revert "update master (#6)" (#17)
This reverts commit ba45a0e6bf2a1e9cc4c57e06462bed7bd097cd07.

+++ src/core/symbology/qgsrenderer.h
@@ -342,7 +342,7 @@ class CORE_EXPORT QgsFeatureRenderer
     virtual QString legendClassificationAttribute() const { return QString(); }
 
     //! Sets type and size of editing vertex markers for subsequent rendering
-    void setVertexMarkerAppearance( int type, double size );
+    void setVertexMarkerAppearance( int type, int size );
 
     /**
      * Returns whether the renderer will render a feature or not.
@@ -505,7 +505,7 @@ class CORE_EXPORT QgsFeatureRenderer
     //! The current type of editing marker
     int mCurrentVertexMarkerType;
     //! The current size of editing marker
-    double mCurrentVertexMarkerSize;
+    int mCurrentVertexMarkerSize;
 
     QgsPaintEffect *mPaintEffect = nullptr;
